pyproject.toml
[project]
name = "vt-native-simulation-suite"
version = "0.1.0"
description = "Headless accelerated simulation harness for Versatile Thermostat against thermal models."
readme = "CLAUDE.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  # Home Assistant core (heavy dependency tree).
  "homeassistant==2025.1.4",
  # Test harness.
  "pytest==8.3.4",
  "pytest-asyncio>=0.24.0",
  "pytest-timeout>=2.3.1",
  "pytest-homeassistant-custom-component==0.13.205",
  # Scenario + analysis.
  "PyYAML>=6.0.1",
  "matplotlib>=3.8,<4",
  "pandas>=2.2,<3",
  "ruamel.yaml>=0.18",
  "fastapi>=0.135.1",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.42.0",
  "httpx>=0.27.2",
  "anyio>=4.12.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
package = false
prerelease = "allow"

[dependency-groups]
dev = [
    "pytest-asyncio>=0.24.0",
]
